changed test folder to split into unit and integration
