+++ tests/unit/test_user_publisher.py
